pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "tnnt-housekeeping"
description = "Terra Nanotech Housekeeping for Alliance Auth"
readme = "README.md"
keywords = [
    "allianceauth",
    "housekeeping",
    "terra_nanotech",
]
license = { file = "LICENSE" }
authors = [
    { name = "Peter Pfeufer", email = "develop@ppfeufer.de" },
]
requires-python = ">=3.10,<3.14"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "allianceauth>=4.10,<5",
]
optional-dependencies.tests-allianceauth-latest = [
    "aa-gdpr",
    "coverage",
    "django-webtest",
]
urls.Changelog = "https://github.com/terra-nanotech/tn-nt-auth-housekeeping/blob/master/CHANGELOG.md"
urls.Codecov = "https://codecov.io/gh/terra-nanotech/tn-nt-auth-housekeeping"
urls.Documentation = "https://github.com/terra-nanotech/tn-nt-auth-housekeeping/blob/master/README.md"
urls.GitHub = "https://github.com/terra-nanotech/tn-nt-auth-housekeeping"

[tool.hatch.build]
include = [
    "/tnnt_housekeeping",
]

[tool.hatch.version]
path = "tnnt_housekeeping/__init__.py"
